LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssVariableReferenceValue-invalid-expected.txt


FAIL Constructing a CSSVariableReferenceValue with an empty variable name throws a TypeError assert_throws_js: function "() => new CSSVariableReferenceValue('')" threw object "ReferenceError: Can't find variable: CSSVariableReferenceValue" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Constructing a CSSVariableReferenceValue with an invalid variable name throws SyntaxError assert_throws_js: function "() => new CSSVariableReferenceValue('bar')" threw object "ReferenceError: Can't find variable: CSSVariableReferenceValue" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
